+++ changelog.md
@@ -24,6 +24,7 @@
 - CEMIFrame: remove `CEMIFrame.cmd`, which can be derived from `CEMIFrame.payload`.
 - Farewell Travis CI; Welcome Github Actions!
 - StateUpdater allow float values for `register_remote_value(tracker_options)` attribute.
+- Handle exceptions from received unsupported or not implemented KNXIP Service Type identifiers
 
 ## 0.15.6 Bugfix for StateUpater 2020-11-26